TodoListリファクタリング
